requirements.txt
Pillow==10.3.0
python-dotenv==1.0.1
streamlit==1.32.2
torch==1.13.0
transformers==4.39.1
